requirements.txt
Flask==3.0.2
requests==2.31.0
firebase-admin==6.4.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
python-dateutil==2.9.0
